refactor(edge/update): move item middleware to middlewares
