+++ api/http/handler/edgeupdateschedules/edgeupdateschedule_update.go
@@ -10,6 +10,7 @@ import (
 	"github.com/portainer/libhttp/request"
 	"github.com/portainer/libhttp/response"
 	portainer "github.com/portainer/portainer/api"
+	"github.com/portainer/portainer/api/http/middlewares"
 )
 
 type updatePayload struct {
@@ -53,7 +54,7 @@ func (payload *updatePayload) Validate(r *http.Request) error {
 // @failure 500
 // @router /edge_update_schedules [post]
 func (handler *Handler) update(w http.ResponseWriter, r *http.Request) *httperror.HandlerError {
-	item, err := FetchItem[portainer.EdgeUpdateSchedule](r)
+	item, err := middlewares.FetchItem[portainer.EdgeUpdateSchedule](r, contextKey)
 	if err != nil {
 		return httperror.InternalServerError(err.Error(), err)
 	}
